tidy up tests
